# api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LanguageLearn API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/courses/import": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Import Course Endpoint",
        "operationId": "import_course_endpoint_v1_courses_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/courses/{course_id}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Course Detail",
        "operationId": "get_course_detail_v1_courses__course_id__get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/courses/{course_id}/concepts": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List Concepts Endpoint",
        "operationId": "list_concepts_endpoint_v1_courses__course_id__concepts_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Course Id"
            }
          },
          {
            "name": "cefr_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CefrLevel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cefr Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConceptSummary"
                  },
                  "title": "Response List Concepts Endpoint V1 Courses  Course Id  Concepts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/concepts/{concept_id}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Concept Detail",
        "operationId": "get_concept_detail_v1_concepts__concept_id__get",
        "parameters": [
          {
            "name": "concept_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Concept Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConceptDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/exercises/submit": {
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Submit Exercise",
        "operationId": "submit_exercise_v1_exercises_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseSubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CefrLevel": {
        "type": "string",
        "enum": [
          "A1",
          "A2",
          "B1",
          "B2",
          "C1",
          "C2"
        ],
        "title": "CefrLevel"
      },
...
      "ConceptDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "concept_type": {
            "$ref": "#/components/schemas/ConceptType"
          },
          "cefr_level": {
            "$ref": "#/components/schemas/CefrLevel"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "source_text": {
            "type": "string",
            "title": "Source Text"
          },
          "target_text": {
            "type": "string",
            "title": "Target Text"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "prerequisites": {
            "items": {
              "$ref": "#/components/schemas/PrerequisiteInfo"
            },
            "type": "array",
            "title": "Prerequisites"
          },
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/ExerciseResponse"
            },
            "type": "array",
            "title": "Exercises"
          }
        },
        "type": "object",
        "required": [
          "id",
          "concept_type",
          "cefr_level",
          "sequence",
          "source_text",
          "target_text",
          "prerequisites",
          "exercises"
        ],
        "title": "ConceptDetail"
      },
      "ConceptImport": {
        "properties": {
          "ref": {
            "type": "string",
            "title": "Ref"
          },
          "concept_type": {
            "$ref": "#/components/schemas/ConceptType"
          },
          "cefr_level": {
            "$ref": "#/components/schemas/CefrLevel"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "source_text": {
            "type": "string",
            "title": "Source Text"
          },
          "target_text": {
            "type": "string",
            "title": "Target Text"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "prerequisites": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prerequisites"
          },
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/ExerciseImport"
            },
            "type": "array",
            "title": "Exercises"
          }
        },
        "type": "object",
        "required": [
          "ref",
          "concept_type",
          "cefr_level",
          "sequence",
          "source_text",
          "target_text"
        ],
        "title": "ConceptImport"
      },
...
      "ConceptSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "concept_type": {
            "$ref": "#/components/schemas/ConceptType"
          },
          "cefr_level": {
            "$ref": "#/components/schemas/CefrLevel"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "source_text": {
            "type": "string",
            "title": "Source Text"
          },
          "target_text": {
            "type": "string",
            "title": "Target Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "concept_type",
          "cefr_level",
          "sequence",
          "source_text",
          "target_text"
        ],
        "title": "ConceptSummary"
      },
      "ConceptType": {
        "type": "string",
        "enum": [
          "vocabulary",
          "grammar"
        ],
        "title": "ConceptType"
      },
      "CourseDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "source_language": {
            "type": "string",
            "title": "Source Language"
          },
          "target_language": {
            "type": "string",
            "title": "Target Language"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "concepts_by_level": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ConceptSummary"
              },
              "type": "array"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/CefrLevel"
            },
            "type": "object",
            "title": "Concepts By Level"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "source_language",
          "target_language",
          "created_at",
          "concepts_by_level"
        ],
        "title": "CourseDetail"
      },
      "CourseImport": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "source_language": {
            "type": "string",
            "title": "Source Language"
          },
          "target_language": {
            "type": "string",
            "title": "Target Language"
          },
          "concepts": {
            "items": {
              "$ref": "#/components/schemas/ConceptImport"
            },
            "type": "array",
            "minItems": 1,
            "title": "Concepts"
          },
          "standalone_exercises": {
            "items": {
              "$ref": "#/components/schemas/ExerciseImport"
            },
            "type": "array",
            "title": "Standalone Exercises"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "title",
          "source_language",
          "target_language",
          "concepts"
        ],
        "title": "CourseImport",
        "description": "Assembled course data (may come from a single JSON or a folder)."
      },
      "CourseImportResponse": {
        "properties": {
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "concepts_created": {
            "type": "integer",
            "title": "Concepts Created"
          },
          "exercises_created": {
            "type": "integer",
            "title": "Exercises Created"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "concepts_created",
          "exercises_created"
        ],
        "title": "CourseImportResponse"
      },
...
      "DependencySource": {
        "type": "string",
        "enum": [
          "manual",
          "auto"
        ],
        "title": "DependencySource"
      },
      "ExerciseImport": {
        "properties": {
          "ref": {
            "type": "string",
            "title": "Ref"
          },
          "exercise_type": {
            "type": "string",
            "title": "Exercise Type"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "concept_refs": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Refs"
          }
        },
        "type": "object",
        "required": [
          "ref",
          "exercise_type"
        ],
        "title": "ExerciseImport"
      },
      "ExerciseResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "exercise_type": {
            "type": "string",
            "title": "Exercise Type"
          },
          "ref": {
            "type": "string",
            "title": "Ref"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "id",
          "exercise_type",
          "ref",
          "data"
        ],
        "title": "ExerciseResponse"
      },
      "ExerciseSubmitRequest": {
        "properties": {
          "concept_id": {
            "type": "string",
            "format": "uuid",
            "title": "Concept Id"
          },
          "exercise_type": {
            "$ref": "#/components/schemas/ExerciseType"
          },
          "user_answer": {
            "type": "string",
            "title": "User Answer"
          },
          "exercise_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exercise Id"
          },
          "review_duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "concept_id",
          "exercise_type",
          "user_answer"
        ],
        "title": "ExerciseSubmitRequest"
      },
      "ExerciseSubmitResponse": {
        "properties": {
          "correct": {
            "type": "boolean",
            "title": "Correct"
          },
          "correct_answer": {
            "type": "string",
            "title": "Correct Answer"
          },
          "normalized_user_answer": {
            "type": "string",
            "title": "Normalized User Answer"
          },
          "new_forward_difficulty": {
            "$ref": "#/components/schemas/ExerciseType"
          },
          "forward_consecutive_correct": {
            "type": "integer",
            "title": "Forward Consecutive Correct"
          },
          "new_reverse_difficulty": {
            "$ref": "#/components/schemas/ExerciseType"
          },
          "reverse_consecutive_correct": {
            "type": "integer",
            "title": "Reverse Consecutive Correct"
          },
          "is_mastered": {
            "type": "boolean",
            "title": "Is Mastered"
          },
          "fsrs_due": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fsrs Due"
          },
          "difficulty_advanced": {
            "type": "boolean",
            "title": "Difficulty Advanced"
          },
          "mastery_changed": {
            "type": "boolean",
            "title": "Mastery Changed"
          }
        },
        "type": "object",
        "required": [
          "correct",
          "correct_answer",
          "normalized_user_answer",
          "new_forward_difficulty",
          "forward_consecutive_correct",
          "new_reverse_difficulty",
          "reverse_consecutive_correct",
          "is_mastered",
          "fsrs_due",
          "difficulty_advanced",
          "mastery_changed"
        ],
        "title": "ExerciseSubmitResponse"
      },
      "ExerciseType": {
        "type": "string",
        "enum": [
          "forward_mc",
          "reverse_mc",
          "cloze",
          "reverse_cloze",
          "forward_typing",
          "reverse_typing"
        ],
        "title": "ExerciseType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrerequisiteInfo": {
        "properties": {
          "concept_id": {
            "type": "string",
            "format": "uuid",
            "title": "Concept Id"
          },
          "source_text": {
            "type": "string",
            "title": "Source Text"
          },
          "target_text": {
            "type": "string",
            "title": "Target Text"
          },
          "cefr_level": {
            "$ref": "#/components/schemas/CefrLevel"
          },
          "source": {
            "$ref": "#/components/schemas/DependencySource"
          }
        },
        "type": "object",
        "required": [
          "concept_id",
          "source_text",
          "target_text",
          "cefr_level",
          "source"
        ],
        "title": "PrerequisiteInfo"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}